requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
amqp
annotated-types
anyio
appdirs
asgiref
asttokens
attrs
backoff
bcrypt
billiard
blinker
build
cachetools
celery
certifi
cffi
charset-normalizer
chromadb
click
click-didyoumean
click-plugins
click-repl
coloredlogs
cryptography
decorator
diskcache
distro
dj-rest-auth
Django
django-allauth
django-cors-headers
django-filter
django-jazzmin
django-sslserver
djangorestframework
djangorestframework_simplejwt
docstring_parser
drf-yasg
durationpy
et_xmlfile
executing
filelock
flatbuffers
frozenlist
fsspec
google-auth
googleapis-common-protos
grpcio
gunicorn
h11
hf-xet
httpcore
httptools
httpx
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
inflection
instructor
ipython
ipython_pygments_lexers
jedi
Jinja2
jiter
json5
json_repair
jsonpickle
jsonref
jsonschema
jsonschema-specifications
kombu
kubernetes
litellm
markdown-it-py
MarkupSafe
matplotlib-inline
mdurl
mmh3
mpmath
multidict
networkx
numpy
oauthlib
onnxruntime
openai
openpyxl
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson
overrides
packaging
parso
pdfminer.six
pdfplumber
pexpect
pillow
portalocker
posthog
prompt_toolkit
propcache
protobuf
psycopg2-binary
ptyprocess
pure_eval
pyasn1
pyasn1_modules
pybase64
pycparser
pydantic
pydantic_core
Pygments
PyJWT
pypdfium2
PyPika
pyproject_hooks
python-dateutil
python-decouple
python-dotenv
pytz
pyvis
PyYAML
redis
referencing
regex
reportlab
requests
requests-oauthlib
rich
rpds-py
rsa
shellingham
six
sniffio
sqlparse
stack-data
sympy
tenacity
tiktoken
tokenizers
tomli
tomli_w
tqdm
traitlets
typer
typing-inspection
typing_extensions
tzdata
uritemplate
urllib3
uv
uvicorn
uvloop
vine
watchfiles
wcwidth
websocket-client
websockets
whitenoise
yarl
zipp
ezdxf
drf-spectacular